+++ app/Http/Controllers/Backend/CountryVariantController.php
@@ -6,6 +6,7 @@ use App\Contracts\CountryVariantInterface;
 use App\Http\Controllers\Controller;
 use App\Models\Backend\CountryVariant;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CountryVariantController extends Controller
 { 
@@ -21,7 +22,7 @@ class CountryVariantController extends Controller
      */
     public function index()
     {
-        
+        Gate::authorize('utility.index');
        $countryVariants = $this->countryVariants->all();  
        return view('backend.country-variant.index', compact('countryVariants'));
     }
@@ -39,6 +40,7 @@ class CountryVariantController extends Controller
      */
     public function store(Request $request)
     { 
+        Gate::authorize('utility.index');
         $this->validate($request,[
             'name'  => 'required'
         ]);
@@ -62,6 +64,7 @@ class CountryVariantController extends Controller
      */
     public function edit(CountryVariant $countryVariant)
     { 
+        Gate::authorize('utility.edit');
         $countryVariant = $countryVariant;
         return view('backend.country-variant.form', compact('countryVariant'));
     }
@@ -71,6 +74,7 @@ class CountryVariantController extends Controller
      */
     public function update(Request $request, CountryVariant $countryVariant)
     { 
+        Gate::authorize('utility.edit');
         $this->validate($request,[
             'name'  => 'required'
         ]);
@@ -85,6 +89,7 @@ class CountryVariantController extends Controller
      */
     public function destroy(CountryVariant $countryVariant)
     { 
+        Gate::authorize('utility.delete');
         if($countryVariant->deletable == 1){
             $countryVariant->delete();
             notify()->success('delete Successfully','Success');
